# swagger.full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PIM API",
    "version": "1.4.1"
  },
  "paths": {
...
    "/api/product/{id}": {
      "get": {
        "description": "Returns product by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Product id to return"
          },
          {
            "name": "attributeCodes",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Attribute codes to aggregate for product"
          },
          {
            "name": "foodUnitAttributeCode",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Attribute code for food unit to use in attribute aggregation"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/product": {
      "get": {
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Accept language header for filtering name translations"
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Zero based page number"
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Number of products per page"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Part of product name for filtering"
          },
          {
            "name": "contributionList",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Product contributions separated with commas"
          },
          {
            "name": "brand",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Product brand"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Product category"
          },
          {
            "name": "quantity",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "Product quantity"
          },
          {
            "name": "measure",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "Product measure such as weight or volume"
          },
          {
            "name": "unit",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Unit of product measure such as kg or l"
          },
          {
            "name": "attributeCodes",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Attribute codes separated with commas to aggregate for product.\n\nFor example, GHG will return carbon footprint.\n\nENERC,FAT,FASAT,CHOAVL,SUGAR,FIBRC,PROT,NACL returns basic nutritional information."
          },
          {
            "name": "foodUnitAttributeCode",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Attribute code for food unit to use in attribute aggregation.\n\nFor example, PORTS is small portion, PORTM is medium portion and PORTL is large portion."
          }
        ],
        "description": "Returns list of products",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Creates a new product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "contributionList": {
            "type": "string",
            "description": "Product contributions separated with comma"
          },
          "measure": {
            "type": "number",
            "description": "Product amount measure"
          },
          "unit": {
            "type": "string",
            "description": "Product amount unit"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductAttribute"
            },
            "description": "List of product attributes"
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductContribution"
            }
          }
        }
      },
      "ProductContribution": {
        "type": "object",
        "required": ["id", "amount", "unit"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "contribution": {
            "$ref": "#/components/schemas/Category"
          }
        }
      },
      "Attribute": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "object",
            "properties": {
              "fi-FI": {
                "type": "string"
              },
              "en-US": {
                "type": "string"
              },
              "sv-SE": {
                "type": "string"
              }
            }
          },
          "parent": {
            "$ref": "#/components/schemas/Attribute"
          },
          "parentId": {
            "type": "integer"
          }
        }
      },
      "ProductAttribute": {
        "type": "object",
        "required": ["id", "name", "value"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "value": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "productId": {
            "type": "integer"
          },
          "attribute": {
            "$ref": "#/components/schemas/Attribute"
          },
          "attributeId": {
            "type": "integer"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "object"
          }
        }
      },
...
    }
  }
}